backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
chromadb
sentence-transformers
openai